+++ dune/stuff/common/fixed_map.hh
@@ -126,7 +126,7 @@ public:
     :map_(map)
   {}
 
-  int getIdx(const key_type& key) const {
+  long getIdx(const key_type& key) const {
     const auto it = std::find_if(map_.begin(), map_.end(), [&](const value_type& val){ return val.first == key;});
     if ( it == map_.end() )
       return -1;
